Hacks in ePub loading.
By default Readium makes HTTP requests to load new content into the
viewer. This requires an HTTP server, something that does not exist
on WinRT (though it could theoretically be built...from scratch).

This hack swaps HTTP requests for script notifications, the XAML
WebViews method of sending messages from the HTML in the view to
the host app. Another way to think of this is Javascript --> C++.

The host application then loads data just like it would in the
standard Readium implementation. Then rather than sending this data
back to the view as an HTTP response like vanilla Readium expects
we instead use WebView's InvokeScriptAsync method (this is the
complement to script notification, or C++ --> Javascript) to send
the data back into the viewer.

At this point the Readium JS code is a tad hacked to facilitate
this new behavior but it can be improved to provide a cleaner API
to interact with.